@@ -2,12 +2,21 @@
 
 namespace App\Controllers;
 
-class AdminController extends BaseController{
+use CodeIgniter\HTTP\RequestInterface;
+use CodeIgniter\HTTP\ResponseInterface;
+use Psr\Log\LoggerInterface;
 
-    public function index()
+class AdminController extends BaseController
+{
+
+    /**
+     * Constructor.
+     */
+    public function initController(RequestInterface $request, ResponseInterface $response, LoggerInterface $logger)
     {
-        if ($this->user == null || $this->user->admin != 1) {
-            return redirect()->to('/connexion');
-        } 
+        parent::initController($request, $response, $logger);
+        if ($this->user->getPrivilege() == null || $this->user->getPrivilege()->getId() < 3) {
+            return redirect()->to('/connection');
+        }
     }
 }
