@@ -1,18 +1,16 @@
 <?php
 
 namespace App\Controllers;
-use App\Models\ProduitModel;
 
 
-class Dashboard extends BaseController
+class Dashboard extends AdminController
 {
 
     public function index()
     {
- 
         $this->data['title'] = "Dashboard";
         $this->data['page'] = "dashboard";
-        $this->data['content'] = view(($this->data['admin']?"admin":"customer").'/dashboard', array());
+        $this->data['content'] = view(($this->data['admin'] ? "admin" : "customer") . '/dashboard', array());
         return view('application', $this->data);
     }
 }
